1. 更新 TreeSaveAllNode & TreeDeleteAllNode 的所拿到的頁面 2. 新增 TreeDeleteAllNode 3. 更新樹 Node 的顏色 4. 增加 SaveAllNode & DeleteAllNode 權限管理

+++ app/controllers/main_controller.rb
@@ -466,34 +466,54 @@ class MainController < ApplicationController
     # 把 Node 存進 Db 裡面，並作一些處理
     #
 	def tree_save_all_node
-        # 把東西存起來
-		node = TreeInfo.where(id: params[:id])[0]
-		node.info = params[:TreeInfo].to_s.gsub("\"","\'")
-		node.save
         
-        # 要先把以前的東西刪掉
-        TreeLink.where(treeinfo_id: params[:id]).destroy_all
-        
-        # 要去 Trace 所有的樹，處理下游議題
-        require 'json'
-        @parseItem = JSON.parse(params[:TreeInfo])
-        @parseItem = [@parseItem['item']]
-        
-        while @parseItem.length != 0
-            childArray = @parseItem[0]['parent']
-            childArray.each do |childItem|
-                item = TreeLink.new
-                item.issue_id = @parseItem[0]['id']
-                item.treeinfo_id = params[:id]
-                item.children_id = childItem['id']
-                item.save
-                @parseItem.push(childItem)
+        # 使用者判斷
+        if current_user != nil
+            if current_user.level >= 2
+                
+                # 把東西存起來
+                node = TreeInfo.where(id: params[:id])[0]
+                node.info = params[:TreeInfo].to_s.gsub("\"","\'")
+                node.save
+
+                # 要先把以前的東西刪掉
+                TreeLink.where(treeinfo_id: params[:id]).destroy_all
+
+                # 要去 Trace 所有的樹，處理下游議題
+                require 'json'
+                @parseItem = JSON.parse(params[:TreeInfo])
+                @parseItem = [@parseItem['item']]
+
+                while @parseItem.length != 0
+                    childArray = @parseItem[0]['parent']
+                    childArray.each do |childItem|
+                        item = TreeLink.new
+                        item.issue_id = @parseItem[0]['id']
+                        item.treeinfo_id = params[:id]
+                        item.children_id = childItem['id']
+                        item.save
+                        @parseItem.push(childItem)
+                    end
+
+                    # 把第一個丟掉
+                    @parseItem.shift
+                end
             end
-            
-            # 把第一個丟掉
-            @parseItem.shift
         end
 	end
+	
+	#
+    # 把 Node 存 Db 刪除
+	#
+    def tree_delete_all_node
+        # 使用者判斷
+        if current_user != nil
+            if current_user.level >= 2
+                TreeInfo.where(id: params[:id]).destroy_all
+            end
+        end
+    end
+
